move CouchDBMRIImporter class to separate file
